auth/requirements.txt
Flask
PyJWT
bcrypt
psycopg2-binary
gunicorn
requests
